Experimenting with a TaskExecutorDistributor that does not use synchronization.  Instead it uses an atomic integer to control access within the within the worker.
It works, but the performance benifits were not like I expected (even when I slightly reduced garuntees in order to allow more inserts).

The goal was to allow in parallel inserts of items into the queue.  Then only have exclusive access when needing to remove or change the queue.  This works, except we spend so much cycles spinning, that any benifits gained are easily lost.  So for, shelving the work as a proof of concept, to maybe pick up later if I get more ideas.