Add a blocking rate limiter implementation
This may be moved to threadly_examples, I don't expect it to be merged right now (at a minimum it would need javadocs and unit tests to be merged).  This is an example of how you can easily make a more guava style rate limiter by using a simple NoThreadScheduler, and our existing rate limiter executor.