Internally use nanos to get more accurate scheduled task execution times
This is to resolve #178, though I am not sure I like it or not yet.
I need to benchmark it and see how it helps.  This does move some math from task execution time to task submission time.  But on the other hand the benefits are not obvious, and it makes other things more complicated.