RateLimiterExecutor - Remove lock during rate calculation
This removes the "rateLock" inside the "RateLimiterExecutor", and instead does optimistic CAS operations.
Internally we are using an "AtomicLong", and doing conversions using Double.doubleToRawLongBits and Double.longBitsToDouble.
We do compare and swap operations until we are able to without contention able to execute the task at the set rate.

Advantages:
  * Obviously no locking
Disadvantages:
  * Requires going to native code each time we call "Double.doubleToLongBits" and "Double.longBitesToDouble"
  * CAS operations may fail requiring the delay calculation to be redone