Extend Pair more to "VolatilePair" and "LazyPair"
This changes the internals of Pair quite a bit.  Adding complexity sadly, but in the effort to make it easier to extend the class.  We now have to access the left/rigth exclusively through getRight() and getLeft().  This avoids needing to extend more functionality in the future if you need to put behavior around getting the values.

VolatilePair is similar to MutablePair, except that it holds volatile references internally.
LazyPair takes in callables to lazily calculate the values, allowing them only produce the values once they need to be retrieved.