ConcurrentArrayList: Avoid synchronization and instead create poor mans lock
This is the revival of an old branch, and old locking scheme.  At the time the performance gains were not there, but now might be different.

The big thing with this is that it still parks threads when they can't apply updates.  In addition there is several native CAS operations.  That may end up killing possible performance gains.