ExecutorLimiter: Attempt to reduce lock contention for consuming tasks by using ReschedulingOperation
By using a ReschedulingOperation (with a SameThreadSubmitterExecutor), under the hood we are using an AtomicInteger so that only one thread is actually consuming the queue.
But we safely are tracking if operations are continuing to happen so that we make sure that the queue is always in a checked state if possible.
This means that threads which can't do this check can just indicate the dirty state then move on, letting the thread return to the parent pool for other uses (perhaps to execute more tasks!).