ExecutorLimiter: Non-locking implementation for task consumption
This optimistically removes from the queue if we have the space in the pool.
If we end up not having a task then we will remove our consumed poll consumption (which results in another native call, as well as the need to retry our queue consume in case threads gave up due to us consuming the slot).

`SchedulerServiceLimiter` ends up overriding the consumption in order to specify the old synchronized implementation.  This synchronization is still needed here to handle task removal.  Because we can not atomically remove from the queue and submit to the pool we must lock for task removal (otherwise we might miss the task in this limbo state).