Possible fix for issue #274
Unit tests are failing because I think the failure shows the issue with this type of fix.
While this seemingly would work it causes several issues:
* The cancel now returns before all in-thread listeners are completed, this introduces a subtle new behavior
* There is more overhead on a cancel on the executing pool

We could in theory track in AbstractCompletableListenableFuture if any optimized listeners were added and reduce these impacts.  However that also has overheads in a larger object size and state checking.